+++ packages/core/src/storage/derived.js
@@ -1,6 +1,6 @@
 import {tick, untrack} from 'svelte';
 import {
-    addDay, addDuration, cloneDate, createView, isFunction, prevClosestDay, subtractDay,
+    addDay, addDuration, cloneDate, createView, isFunction, prevClosestDay, setMidnight, subtractDay,
     toEventWithLocalDates, toViewWithLocalDates
 } from '#lib';
 
@@ -91,8 +91,9 @@ export function viewDates(mainState) {
         let dates = [];
 
         untrack(() => {
-            let date = cloneDate(activeRange.start);
-            let end = cloneDate(activeRange.end);
+            // activeRange may be offset by hours due to slotMaxTime, so we set it to midnight
+            let date = setMidnight(cloneDate(activeRange.start));
+            let end = setMidnight(cloneDate(activeRange.end));
             while (date < end) {
                 if (!hiddenDays.includes(date.getUTCDay())) {
                     dates.push(cloneDate(date));